% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_wb_data.R
\name{pivot_wb_data}
\alias{pivot_wb_data}
\title{Pivot World Bank data}
\usage{
pivot_wb_data(data.name, var.name)
}
\arguments{
\item{data.name}{Name of the original data set}

\item{var.name}{Name of the new variable}
}
\description{
Pivot World Bank data
}
\section{Functions}{
\itemize{
\item \code{pivot_wb_data()}: function turns wide World Bank data into a long data format

}}
\examples{
# Use pivot_wb_data() on maternal mortality data
pivot_wb_data(maternal_mortality, "maternal_mortality")
}
\author{
Aya Mitani
}
